+++ components/src/main/kotlin/by/jprof/telegram/components/config/dynamo.kt
@@ -1,6 +1,5 @@
 package by.jprof.telegram.components.config
 
-import org.koin.core.qualifier.named
 import org.koin.dsl.module
 import software.amazon.awssdk.services.dynamodb.DynamoDbAsyncClient
