+++ webhook/src/main/kotlin/by/jprof/telegram/opinions/webhook/config/dynamo.kt
@@ -4,13 +4,8 @@ import by.jprof.telegram.opinions.webhook.dao.KeyboardsDAO
 import by.jprof.telegram.opinions.webhook.dao.KotlinMentionsDAO
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
-import software.amazon.awssdk.services.dynamodb.DynamoDbAsyncClient
 
 val dynamoModule = module {
-    single {
-        DynamoDbAsyncClient.create()
-    }
-
     single {
         KotlinMentionsDAO(
                 get(),
