+++ youtube/src/main/kotlin/by/jprof/telegram/opinions/youtube/config/env.kt
@@ -3,13 +3,13 @@ package by.jprof.telegram.opinions.youtube.config
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
-const val TABLE_VOTES = "TABLE_VOTES"
 const val YOUTUBE_API_TOKEN = "YOUTUBE_API_TOKEN"
+const val TABLE_YOUTUBE_CHANNELS_WHITELIST = "TABLE_YOUTUBE_CHANNELS_WHITELIST"
 
 val youtubeEnvModule = module {
     listOf(
-        TABLE_VOTES,
-        YOUTUBE_API_TOKEN
+        YOUTUBE_API_TOKEN,
+        TABLE_YOUTUBE_CHANNELS_WHITELIST
     ).forEach { variable ->
         single(named(variable)) {
             System.getenv(variable)!!
